requirements.txt
tensorflow==2.15.0
opencv-python
numpy
matplotlib
tqdm
scikit-learn
pillow
